src/requirements.txt
asgiref==3.3.1
astroid==2.5
autopep8==1.5.5
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Django==3.1.6
google-api-core==1.26.0
google-auth==1.27.0
google-cloud-core==1.6.0
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mysqlclient==2.0.3
packaging==20.9
Pillow==8.1.0
protobuf==3.15.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
spotipy==2.16.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.3
wrapt==1.12.1
